requirements.txt
requests
mysql-connector-python
colorama
datetime
discord_webhook